auth-commit

+++ backend/controllers/goalController.js
@@ -2,9 +2,10 @@
 //@route GET /api/goals
 //@access Private
 const Goal = require('../model/goalmodel');
+const User = require('../model/userModel');
 const asyncHandler = require('express-async-handler');
 const getGoals = asyncHandler(async (req,res)=>{
-    const goals = await Goal.find();
+    const goals = await Goal.find({user:req.user.id});
     res.status(200).send(goals);
 });
 const setGoals = asyncHandler(async (req,res)=>{
@@ -13,7 +14,8 @@ const setGoals = asyncHandler(async (req,res)=>{
         throw new Error('Please add a text field')
 }
 const goal = await Goal.create({
-    text:req.body.text
+    text:req.body.text,
+    user:req.user.id
 });
     res.status(200).send(goal);
 });
@@ -22,7 +24,17 @@ const UpdateGoals = asyncHandler(async (req,res)=>{
     if(!goal){
         res.status(400)
         throw new Error('Goal not found');
-}
+    }
+    const user = await User.findById(req.user.id);
+    if(!user){
+        res.status(401)
+        throw new Error("User not found");
+    }
+    //make sure the logged in user matches the goal user
+    if(goal.user.toString() !== user.id){
+        res.status(401);
+        throw new Error("User not authorized");
+    }
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id,req.body,{new: true});
     res.status(200).send(updatedGoal);
 });
@@ -31,6 +43,16 @@ const DeleteGoals = asyncHandler(async (req,res)=>{
     if(!goal){
         res.status(400)
         throw new Error('Goal not found');
+    }
+     const user = await User.findById(req.user.id);
+    if(!user){
+        res.status(401)
+        throw new Error("User not found");
+    }
+    //make sure the logged in user matches the goal user
+    if(goal.user.toString() !== user.id){
+        res.status(401);
+        throw new Error("User not authorized");
     }
     await Goal.findByIdAndDelete(req.params.id);
     res.status(200).send({id:req.params.id});
